chore: move analytics to general
