+++ nodejs/src/ingestion/general/heatmap-subpipeline.ts
