+++ nodejs/src/ingestion/general/joined-ingestion-pipeline.ts
