+++ nodejs/src/ingestion/general/registry.ts
